Remove mock file
