+++ pkg/cardtoken/client_test.go
@@ -9,6 +9,7 @@ import (
 	"reflect"
 	"strings"
 	"testing"
+	"time"
 
 	"github.com/mercadopago/sdk-go/pkg/config"
 	"github.com/mercadopago/sdk-go/pkg/internal/httpclient"
@@ -92,4 +93,34 @@ func TestCreate(t *testing.T) {
 			}
 		})
 	}
+}
+
+func mockCardToken() *Response {
+	return &Response{
+		ID:              "3d40b34eb41a6d0923e5bc545927c2e9",
+		FirstSixDigits:  "503143",
+		ExpirationMonth: 11,
+		ExpirationYear:  2025,
+		LastFourDigits:  "6351",
+		Cardholder: CardholderResponse{
+			Identification: IdentificationResponse{
+				Number: "70383868084",
+				Type:   "CPF",
+			},
+			Name: "MASTER TEST",
+		},
+		Status:             "active",
+		DateCreated:        parseDate("2024-02-08T09:05:42.725-04:00"),
+		DateLastUpdated:    parseDate("2024-02-08T09:05:42.725-04:00"),
+		DateDue:            parseDate("2024-02-16T09:05:42.725-04:00"),
+		LuhnValidation:     true,
+		LiveMode:           false,
+		CardNumberLength:   16,
+		SecurityCodeLength: 3,
+	}
+}
+
+func parseDate(s string) *time.Time {
+	d, _ := time.Parse(time.RFC3339, s)
+	return &d
 }
